refactor and break into modules
