+++ middleware/static.js
@@ -1,0 +1,4 @@
+module.exports = function initializeStatic (app, express) {
+  app.use(express.static('static'));
+  app.use('/static', express.static('static'));
+};
