+++ models/connectDB.js
@@ -1,0 +1,5 @@
+const db = require('./dataBase');
+
+module.exports = function connectDB () {
+  db.sync();
+};
