+++ routers/add-albums.js
@@ -1,0 +1,8 @@
+function addAlbums (app) {
+  app.get('/add-albums', (req, res) => {
+    res.render('add-albums');
+  });
+}
+
+module.exports = addAlbums;
+
