+++ routers/logout.js
@@ -1,0 +1,8 @@
+function logOut (app) {
+  app.get('/logout',(req, res) => {
+    req.session = null;
+    res.redirect('/');
+  });
+}
+
+module.exports = logOut;
