+++ strategy/googleStrategy.js
@@ -1,0 +1,21 @@
+const passport = require('passport');
+const GoogleStrategy = require('passport-google-oauth2').Strategy;
+require('dotenv').config();
+
+module.exports = function googleStrategy () {
+  passport.serializeUser((user, done) => {
+    done(null, user);
+  });
+  passport.deserializeUser((user, done) => {
+    done(null, user);
+  });
+  
+  passport.use(new GoogleStrategy({
+    clientID: process.env.clientID,
+    clientSecret: process.env.clientSecret,
+    callbackURL: process.env.callbackURL
+  },
+  function(accessToken, refreshToken, profile, done) {
+    return done(null, profile);
+  }));
+};
